pyproject.toml
[tool.poetry]
name = "gpt-translation-api"
version = "0.1.0"
description = ""
authors = ["Silas Nordgren <nordgrensilas@gmail.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "gpt_translation_api"}]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.95.0"
openai = "^0.27.4"
pydantic = "^1.10.7"
uvicorn = "^0.21.1"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
